Refactor:  rename `actions/load-rubric` to `actions/rubric`
PT Story:
If a teacher uses a rubric to give feedback, it is copied and locked for them for that assignment, so that it does not change if the author changes it after that point.

[#155368181]

https://www.pivotaltracker.com/story/show/155368181
