+++ js/api-middleware.js
@@ -1,5 +1,5 @@
 import { fetchReportData, updateReportSettings, APIError } from './api'
-import { loadRubric } from './actions/load-rubric'
+import { loadRubric } from './actions/rubric'
 
 // This middleware is executed only if action includes .callAPI object.
 // It calls API action defined in callAPI.type.
